FEATURE: reactions  post serializer
This is not finished draft.

Problem is that this will generate N+1 query when we try to serialize a topic with several posts.
I need to find a way to hook to TopicQuery and `includes(reactions: [:users])` when plugin is enabled.

Nonetheless, this is an idea of how can we enhance post serializer.
I think that because we have all necessary information in PostSerializer we can drop that controller CustomReactionsController (at least index action)